// BettingPlatform.WebApp/src/app/api/model/offerOutcomeDto.ts
/**
 * BettingPlatform.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OfferOutcomeDto { 
    outcomeTemplateId?: string;
    outcomeCode?: string | null;
    odds?: number;
    isEnabled?: boolean;
}